Lift the in-memory BSA builder out of the extract probe
